requirements.txt
amqp==5.1.1
aniso8601==9.0.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dataclasses-json==0.5.7
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
marshmallow==3.16.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.29
pyparsing==3.0.9
pytz==2022.1
requests==2.28.0
six==1.16.0
smartcrop==0.3.3
typing-extensions==4.2.0
typing-inspect==0.7.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.8.0
